Changed ArticleView.kt likes field type to Int.

+++ app/src/main/java/com/example/nattramn/core/utils/Utils.kt
@@ -65,7 +65,7 @@ fun toArticleView(
         date = articleEntity!!.date,
         title = articleEntity.title,
         body = articleEntity.body,
-        likes = articleEntity.likes,
+        likes = articleEntity.likes.toInt(),
         tags = tagsEntity!!.map { it.toTagString() },
         commentViews = commentsEntity?.map { it.toCommentView() }!!,
         commentsNumber = commentsEntity.size,
